% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_candlestick}
\alias{plot_candlestick}
\title{#' Plot candlestick with SMAs}
\usage{
plot_candlestick(data)
}
\arguments{
\item{spy}{SPY data frame with OHLC and SMA columns}
}
\value{
ggplot object
}
\description{
#' Plot candlestick with SMAs
}
