% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rsi}
\alias{plot_rsi}
\title{Plot RSI with overbought and oversold thresholds}
\usage{
plot_rsi(
  data,
  rsi_col = "RSI_14",
  oversold = 30,
  overbought = 70,
  colors = candle_palette(),
  date_breaks = "3 month",
  date_labels = "\%b \%y"
)
}
\arguments{
\item{data}{SPY data frame with a date and RSI column}

\item{rsi_col}{Name of the RSI column to plot}

\item{oversold}{Lower threshold line}

\item{overbought}{Upper threshold line}

\item{colors}{Candle up/down colors, see \code{\link[=candle_palette]{candle_palette()}}}

\item{date_breaks}{Date break spacing}

\item{date_labels}{Date label format}
}
\value{
ggplot object
}
\description{
Plot RSI with overbought and oversold thresholds
}
