% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_volume}
\alias{plot_volume}
\title{Plot volume below candlestick chart}
\usage{
plot_volume(
  data,
  colors = candle_palette(),
  date_breaks = "3 month",
  date_labels = "\%b \%y"
)
}
\arguments{
\item{data}{SPY data frame with date, open, close and volume columns}

\item{colors}{Candle up/down colors, see \code{\link[=candle_palette]{candle_palette()}}}

\item{date_breaks}{Date break spacing}

\item{date_labels}{Date label format}
}
\value{
ggplot object
}
\description{
Plot volume below candlestick chart
}
